+++ modules/Students/Menu.php
@@ -27,9 +27,7 @@ $menu['Students']['admin'] = array(
 	'Students/PrintStudentInfo.php' => _( 'Print Student Info' ),
 	2 => _( 'Setup' ),
 	'Students/StudentFields.php' => _( 'Student Fields' ),
-	'Students/AddressFields.php' => _( 'Address Fields' ),
-	'Students/PeopleFields.php' => _( 'Contact Fields' ),
-	'Students/EnrollmentCodes.php' => _( 'Enrollment Codes' )
+	'Students/EnrollmentCodes.php' => _( 'Enrollment Codes' ),
 );
 
 $menu['Students']['teacher'] = array(
@@ -40,16 +38,16 @@ $menu['Students']['teacher'] = array(
 	1 => _( 'Reports' ),
 	'Students/AdvancedReport.php' => _( 'Advanced Report' ),
 	'Students/StudentLabels.php' => _( 'Print Student Labels' ),
-	'Students/Letters.php' => _( 'Print Letters' )
+	'Students/Letters.php' => _( 'Print Letters' ),
 );
 
 $menu['Students']['parent'] = array(
 	'title' => _( 'Students' ),
 	'default' => 'Students/Student.php',
-	'Students/Student.php' => _( 'Student Info' )
+	'Students/Student.php' => _( 'Student Info' ),
 );
 
 $exceptions['Students'] = array(
 	'Students/Student.php&include=General_Info&student_id=new' => true,
-	'Students/AssignOtherInfo.php' => true
+	'Students/AssignOtherInfo.php' => true,
 );