+++ modules/Students/includes/PeopleFields.php
@@ -1,16 +1,12 @@
 <?php
 /**
- * Address Fields
+ * People Fields
  *
  * @package RosarioSIS
  * @subpackage modules
  */
 
-require_once 'ProgramFunctions/Fields.fnc.php';
-
 DrawHeader( ProgramTitle() );
-
-//$_ROSARIO['allow_edit'] = true;
 
 if ( isset( $_POST['tables'] )
 	&& is_array( $_POST['tables'] )
@@ -56,7 +52,7 @@ if ( isset( $_POST['tables'] )
 					$sql = 'INSERT INTO ' . $table . ' ';
 
 					// New Field.
-					if ( $table === 'ADDRESS_FIELDS' )
+					if ( $table === 'PEOPLE_FIELDS' )
 					{
 						if ( isset( $columns['CATEGORY_ID'] ) )
 						{
@@ -65,16 +61,16 @@ if ( isset( $_POST['tables'] )
 							unset( $columns['CATEGORY_ID'] );
 						}
 
-						$_REQUEST['id'] = AddDBField( 'ADDRESS', 'address_fields_seq', $columns['TYPE'] );
+						$_REQUEST['id'] = AddDBField( 'PEOPLE', 'people_fields_seq', $columns['TYPE'] );
 
 						$fields = 'ID,CATEGORY_ID,';
 
 						$values = $_REQUEST['id'] . ",'" . $_REQUEST['category_id'] . "',";
 					}
 					// New Category.
-					elseif ( $table === 'ADDRESS_FIELD_CATEGORIES' )
+					elseif ( $table === 'PEOPLE_FIELD_CATEGORIES' )
 					{
-						$id = DBSeqNextID( 'ADDRESS_FIELD_CATEGORIES_SEQ' );
+						$id = DBSeqNextID( 'PEOPLE_FIELD_CATEGORIES_SEQ' );
 
 						$fields = "ID,";
 
@@ -123,9 +119,9 @@ if ( $_REQUEST['modfunc'] === 'delete'
 	if ( isset( $_REQUEST['id'] )
 		&& intval( $_REQUEST['id'] ) > 0 )
 	{
-		if ( DeletePrompt( _( 'Address Field' ) ) )
+		if ( DeletePrompt( _( 'Contact Field' ) ) )
 		{
-			DeleteDBField( 'ADDRESS', $_REQUEST['id'] );
+			DeleteDBField( 'PEOPLE', $_REQUEST['id'] );
 
 			// Unset modfunc & ID & redirect URL.
 			RedirectURL( array( 'modfunc', 'id' ) );
@@ -134,10 +130,10 @@ if ( $_REQUEST['modfunc'] === 'delete'
 	elseif ( isset( $_REQUEST['category_id'] )
 		&& intval( $_REQUEST['category_id'] ) > 0 )
 	{
-		if ( DeletePrompt( _( 'Address Field Category' ) . ' ' .
+		if ( DeletePrompt( _( 'Contact Field Category' ) . ' ' .
 				_( 'and all fields in the category' ) ) )
 		{
-			DeleteDBFieldCategory( 'ADDRESS', $_REQUEST['category_id'] );
+			DeleteDBFieldCategory( 'PEOPLE', $_REQUEST['category_id'] );
 
 			// Unset modfunc & category ID & redirect URL.
 			RedirectURL( array( 'modfunc', 'category_id' ) );
@@ -149,6 +145,9 @@ if ( ! $_REQUEST['modfunc'] )
 {
 	echo ErrorMessage( $error );
 
+	// Category menu: student|address|contact.
+	DrawHeader( _fieldsCategoryMenu( $_REQUEST['category'] ) );
+
 	// ADDING & EDITING FORM.
 	if ( $_REQUEST['id']
 		&& $_REQUEST['id'] !== 'new' )
@@ -156,9 +155,9 @@ if ( ! $_REQUEST['modfunc'] )
 		$RET = DBGet( "SELECT ID,CATEGORY_ID,TITLE,TYPE,SELECT_OPTIONS,
 			DEFAULT_SELECTION,SORT_ORDER,REQUIRED,
 			(SELECT TITLE
-				FROM ADDRESS_FIELD_CATEGORIES
+				FROM PEOPLE_FIELD_CATEGORIES
 				WHERE ID=CATEGORY_ID) AS CATEGORY_TITLE
-			FROM ADDRESS_FIELDS
+			FROM PEOPLE_FIELDS
 			WHERE ID='" . $_REQUEST['id'] . "'" );
 
 		$RET = $RET[1];
@@ -169,8 +168,8 @@ if ( ! $_REQUEST['modfunc'] )
 		&& $_REQUEST['category_id'] !== 'new'
 		&& $_REQUEST['id'] !== 'new' )
 	{
-		$RET = DBGet( "SELECT ID AS CATEGORY_ID,TITLE,RESIDENCE,MAILING,BUS,SORT_ORDER
-			FROM ADDRESS_FIELD_CATEGORIES
+		$RET = DBGet( "SELECT ID AS CATEGORY_ID,TITLE,CUSTODY,EMERGENCY,SORT_ORDER
+			FROM PEOPLE_FIELD_CATEGORIES
 			WHERE ID='" . $_REQUEST['category_id'] . "'" );
 
 		$RET = $RET[1];
@@ -179,7 +178,7 @@ if ( ! $_REQUEST['modfunc'] )
 	}
 	elseif ( $_REQUEST['id'] === 'new' )
 	{
-		$title = _( 'New Address Field' );
+		$title = _( 'New Contact Field' );
 
 		$RET['ID'] = 'new';
 
@@ -187,7 +186,7 @@ if ( ! $_REQUEST['modfunc'] )
 	}
 	elseif ( $_REQUEST['category_id'] === 'new' )
 	{
-		$title = _( 'New Address Field Category' );
+		$title = _( 'New Contact Field Category' );
 
 		$RET['CATEGORY_ID'] = 'new';
 	}
@@ -198,34 +197,25 @@ if ( ! $_REQUEST['modfunc'] )
 		$extra_fields = array(
 			'<table class="width-100p cellspacing-0"><tr class="st"><td>' .
 			CheckboxInput(
-				$RET['RESIDENCE'],
-				'tables[' . $_REQUEST['category_id'] . '][RESIDENCE]',
-				_( 'Residence' ),
+				$RET['CUSTODY'],
+				'tables[' . $_REQUEST['category_id'] . '][CUSTODY]',
+				_( 'Custody' ),
 				'',
 				$_REQUEST['category_id'] === 'new',
 				button( 'check' ),
 				button( 'x' )
 			) . '</td><td>' .
 			CheckboxInput(
-				$RET['MAILING'],
-				'tables[' . $_REQUEST['category_id'] . '][MAILING]',
-				_( 'Mailing' ),
-				'',
-				$_REQUEST['category_id'] === 'new',
-				button( 'check' ),
-				button( 'x' )
-			) . '</td><td>' .
-			CheckboxInput(
-				$RET['BUS'],
-				'tables[' . $_REQUEST['category_id'] . '][BUS]',
-				_( 'Bus' ),
+				$RET['EMERGENCY'],
+				'tables[' . $_REQUEST['category_id'] . '][EMERGENCY]',
+				_( 'Emergency' ),
 				'',
 				$_REQUEST['category_id'] === 'new',
 				button( 'check' ),
 				button( 'x' )
 			) . '</td></tr></table>' .
 			FormatInputTitle(
-				_( 'Note: All unchecked means applies to all addresses' ),
+				_( 'Note: All unchecked means applies to all contacts' ),
 				'',
 				false,
 				''
@@ -234,7 +224,7 @@ if ( ! $_REQUEST['modfunc'] )
 	}
 
 	echo GetFieldsForm(
-		'ADDRESS',
+		'PEOPLE',
 		$title,
 		$RET,
 		isset( $extra_fields ) ? $extra_fields : array()
@@ -242,7 +232,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 	// CATEGORIES.
 	$categories_RET = DBGet( "SELECT ID,TITLE,SORT_ORDER
-		FROM ADDRESS_FIELD_CATEGORIES
+		FROM PEOPLE_FIELD_CATEGORIES
 		ORDER BY SORT_ORDER,TITLE" );
 
 	// DISPLAY THE MENU.
@@ -258,7 +248,7 @@ if ( ! $_REQUEST['modfunc'] )
 		&& $categories_RET )
 	{
 		$fields_RET = DBGet( "SELECT ID,TITLE,TYPE,SORT_ORDER
-			FROM ADDRESS_FIELDS
+			FROM PEOPLE_FIELDS
 			WHERE CATEGORY_ID='" . $_REQUEST['category_id'] . "'
 			ORDER BY SORT_ORDER,TITLE", array( 'TYPE' => 'MakeFieldType' ) );